Add a modified version of the McTernan unwinder and use it by default on ARM.
The McTernan unwinder [1] is an unwinder that uses abstract interpretation to
do a reasonable job of unwinding in a hostile environment (where there may be
no frame pointers, DWARF info, asynchronous unwind tables, and so forth). Most
ARM devices in the wild have libraries that fall into at least these categories
(e.g. binary blobs for graphics drivers, prebuilt versions of `libc`), so this
kind of best-effort unwinding turns out to be useful in practice.

I added partial support for Thumb-2 to the McTernan unwinder--enough to get
reasonable stack traces--as well as hacky support for Linux signal handlers and
Rust `__morestack` prologues.

Additionally, this commit changes `pprof` to avoid using `addr2line` on ARM, as
it seems to get confused by the `$a` and `$t` symbols for distinguishing ARM
and Thumb instruction sets.

[1]: http://www.mcternan.me.uk/ArmStackUnwinding/